Adds jsonification of nested dynamic form
Nested dynamic forms require a lot of JS fiddeling to properly
autopost the formdata to the server (updating input element
name indices on add/delete, ...)

Therefore we just construct a quasi in-toto conformant layout
on client side and post the JSON data instead of the formdata
to the server where we construct the actual in-toto layout.

- Adds back a button to add new rules:
We used to listen for the change event in the rule select box, but
there is no straight forward way to trigger the event for the same
selection more than once in a row.

- Makes "PRODUCTS" the default selection for match rule destination
type, to always have valid rules.