Adds layout pubkey select listener to update step
The layout pubkey multiselect field lists the pubkeys currently
available in the session directory.
Only those pubkeys that are selected should be available in the
step pubkey multiselect field.

Therefore we need to update the step pubkey multiselect fields
if a key is selected or deselected in the layout multiselect field.