Adds cmd input field + generalizes snippet update
We now have two input fields for every step, one to specify the
step name and one or the command that should be wrapped.

This commit generalizes the snippet update JS to update a
span identified by a target attribute in the same wiz-step div,
when changing the text in one of the input fields.