Modifies layout wizard views
Adds route with layout name for view layout_wizard to serve an
existing layout, currently "wizarded".

Adds view layout_wizard_generate to generate a new layout based
on the step command form data and uploaded link files.
The uploaded link files and the steps created in the wizard must
be equal in terms of amount and names.