In-toto layout wizard clean slate + flask scaffold
This orphan branch marks the the clean slate for a new in-toto
layout creation wizard, which will be based on UI mockups:
https://github.com/in-toto/layout-web-tool/tree/editor-and-wizard-wip/mockups

Isolated snippets and ideas from the pure layout editor (WIP)
https://github.com/in-toto/layout-web-tool/tree/develop
and the basic wizard plus editor (WIP)
https://github.com/in-toto/layout-web-tool/tree/editor-and-wizard-wip
might end up in this branch.

If we want to have a single flask app with a wizard and a editor
module, we should consider using blueprints:
http://flask.pocoo.org/docs/0.12/blueprints/

This commit provides some basic Flask web scaffolding.