Finishes basic versioning page UI
Basic versioning page consists of a grid for predefined options
(currently dummy options) with radio buttons and a form (input
field) that expands below the option.
Additionally, there is a button to add a custom VCS commands
(currently not limited to one - this should probably be changed).

Expand/collapse and show/hide is performed using custom JS.

The look and feel for this page probably won't change a lot but
the code structure could be more elegant, (e.g. not one .row
per "row").

Also I went back to style using bootstrap classes in the HTML
instead of extending custom classes in SCSS - while this makes
the HTML more convoluted, it is easier for me during prototyping
(trying out bootstrap classes).