Changes code structure for option grid
As announced in 433f469bb1c5acc2ac90bdc014d50f80d2125b59
this commit cleans up the code structure of the options grid.
There are no visible changes in the UI but the HTML and CSS looks
like it should now.

Changes:
All cells are in a single `.row` div and the amount of cells
actually visible on a single row depends on an SCSS constant
(one per screen width breakpoint).

Furthermore, each form is a direct child of an option cell and the
effect of showing it in the subsequent is controlled using SCSS
(above all media queries).

Kudos to mix(in)-master-@pooledge and his awesome styling skills.