Adds gulpfile.js for vendor js copying
All required vendor js files are installed to a local node_modules
dir when running `npm install` in the root of the project.

Instead of serving those assets directly from node_modules (or
from a CDN as we used to), we use gulp.js[1] to copy them to a place
where flask expects, i.e. the ./static/ dir.

Note: Later we could add `gulp` tasks to concatenate and/or minify
or uglify the assets on copy.

Furthermore, this commit
- adds installation instructions and development tips to README.md
- and adds a html5 drag and drop sorting JS library to package.json

[1] http://gulpjs.com/