Adds finished quality management page template
The QA page is similarly structured like VCS and building
(without a sorting section), with a grid of predefined QA options
with expandable forms and a section to add custom commands.

Since the form is more complex than in the previous pages this
commit also introduces a page internal macro for the form for
DRY in predefined and custom option forms.