Adds D3.js, plus basic supply chain page template
Adds D3.js as npm dependency and updates gulpfile to copy it into
vendor dir on `gulp`.

D3.js will be used to render in-toto layouts as graphs, a.o.
on the software supply chain page.

This commit also reads a demo layout (Note: layout not checked
into VCS) in the `software_supply_chain` view and passes it as
json string to the template, where it's assigned to a global
JS variable later used to init the D3 SVG.`