Adds dagre-d3 for DAG layouts plus basic example
dagre-d3 creates layouts for directed acyclic graphs (DAGs) and
renders them using D3. Unfortunately the project is not maintained
but it seems to work pretty well.
dagre-d3 required a downgrade to D3 v3.

An interesting maintained alternative could be
http://js.cytoscape.org/

This commit also adds code (based on dagre examples) to render an
in-toto supply chain.