Updates JS/CSS to draw compound layout graph
in-toto steps and inspections are shown as isolated nodes
(degree == 0) having each a material and a product child node, in
case the node's materials or products are matched against another
node's materials or products.

This commit also removes the client side layout-to-graph data
transformation and uses the server side transformed data instead.