Adds SVG viewBox JS init + basic styling
The viewBox is now initialized to the actual width/height of its container.
This makes the transformation between coordinate systems introduced in
1334ecfed5ad7b46fd23c711c0ebad886d271930 obsolete.

As a consequence dagre-d3 intern translation (e.g. for label positioning) also
work properly now.
Caveat: When resize the window, the svg container does not keep the aspect
ratio but the SVG does. This can lead to cropping when zooming after a window
resize.

This commit also introduces some basic styling
- through element re-ordering, and
- scss