Moves graph drawing to client-side JS in ssc html
The graph should not always be called on window load, but just
on the software supply chain page.

For jQuery convenience in client-side JS (inside script tags) we have to
load it before the body loads.

Later we could factor out this functionality into a js file that
only loads on the ssc page.