Rename and update `option_custom` macro (WIP)
The option_custom macro will actually contain all the form posted
input elements and is now called `form_container`.
The form groups rendered in the form_container are generated from
the passed user_data, which is a dictionary of format
{items: [<posted form data per tool>, ]}

This commit also replaces kwargs with regular arguments and
renames/adds a couple of arguments.

A CSS id is also renamed.

WIP: Templates using that macro need to be adjusted