Add copy option form to form container click listener
According to a recent UI design change, only form content in the
form container gets posted to the server.

If a user edits provided tool defaults in the option grid, he/she
must clone and add that form to the form container below in order
to send the data to the server.

That reasoning is that like this it is more transparent which
"steps" a user actually sends to the server, especially if it is
a mix of provided default commands and custom commands.
Also sorting should be easier like this.

This commit adds the necessary JS to copy the forms.

Thanks to @awwad for his usability input