Add FIXME comment for qa step order to ssc view
The QA form allows to specify if a qa step is run before or after a
certain build step.
The added FIXME comment acknowledges that this information is not
regarded as of now (i.e. no special re-sorting of qa steps happens)

I wonder if we should maybe remove this feature altogether because
re-ordering of steps can be done in the ssc view anyway.