Update ssc template and implement backend
The software supply chain view generates a software supply chain
graph based on previously posted data (vcs, build, qa, package).
Inspections are based on the information from the qa page.

If the view receives a POST request it creates a new graph based
on the posted form data from the ssc template, which is used
subsequently.

There are still a couple of issues with this view/page, which are
listed in the view's docstring.