Fix wrong order bug in chaining and wrap up
The snippets in chaining and wrap up template used to display
steps based on the data in the authorization dict, which (since
it is a dict) does not preserve the order of the steps.

This fixes the bug by using the steps from the "ssc" nodes list.
Downside, we now have to also check if the nodes are of type
"step" (and not "inspection") in the template.