Change opt-form to optionally add and sumit
So a user had to click on an option in the opt form, add it to
the workflow and then click on next to post the form and get to
the next page.
This commit adds an additional button to each opt-form that let's
the user immedeately post the form and continue on the next page.