Refactor software supply chain data structure
The ssc subdocument was tailored for the graph visualization JS
function. This commit simplifies the data structure and the
functions that generate the data structure, either from data
posted in previous steps, or from data posted on the software
supply chain page.

The data in ssc is essential for the whole tool -- it is first
generated using the data from early steps and it is used to
render pages in later steps and to eventually generate a layout
-- this commit requires a lot of changes all over the place.

This commit also temporarily disables graph visualization.