Add modifying/reporting step handling to ssc graph
Reporting steps, e.g. QA steps, should not have an outdegree in the
ssc graph, because they don't have products.

This commit adds a checkbox to the steps in the ssc form and
modifies the graph generating function to ignore reporting
steps in terms of edge sources.

Note:
Checkboxes are only posted to the server if they are checked, but
we need a value per step to be able to aggregate step values by
index, hence we employ a JS workaround that copies the checkbox
state to a hidden input field on change.