Enable link upload using tar(.gz) archive
This commit adds a tar archiving command to the snippet on
the chaining page and changes the chaining view to additionally
accept archives of links.

The commit also updates the link upload dropzone callback, to
display previews for the actually stored files, i.e.:
if foo.tar contains bar.link and baz.link, the dropzone
displays the latter two on successful return from the server.