Adapt `session_to_ssc` to use checkbox data
This commit changes the ssc creation function to use a mix
of user posted data (checkboxes) and data from tooldb.

Since we removed the inspection forms on the qa page, we have to
think of a new way to auto generate inspections.
This commit removes the part in `session_to_ssc` that used to
generate inspections based on form posted data.