Update to in-toto 0.1.1
This version of in-toto uses a new metadata format, where links
and layouts are embedded in a Metablock's "signed" field,
the Metablock also contains a signature field that can carry
signatures over the contents of the signed field.

This commit updates
- requirements.txt to install in-toto from pypi at 0.1.1
- the link loading calls (to handle the new metadata format)
- the layout dumping call (to handle the new metadata format)
- the installation instruction on start page

Also fixes an import that was broken after a module rename.