Update frontend dependencies
Includes updating sources package.json and our frontend task manager
(gulp) that takes care of vendorization and the corresponding html
tags.
The commit also adopts the gulpfile to return a stream in the task
which is required in the new gulp version.

tether.io dependency is dropped because it was only required by
bootstrap 4 alpha, which is stable now.
Bootstrap would require popper.js for tooltips, but since we don't
use tooltips and there are no complaints in the browser's console
I don't include popper for now.